requirements.txt
psycopg2-binary>=2.9
pyyaml>=6.0
resend>=2.0
apscheduler>=3.10
